+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 
 ### Added
-
+- `ar-arca-v4`: Added initial Argentina ARCA addon.
 - `eu-en16931-v2017`: Added check for negative prices.
 - `br-nfse-v1`: Add extensions for IBS/CBS tax determination (RTC)
 - `br`: Add CBS and IBS tax categories